+++ packages/leancode_debug_page/example/lib/main.dart
@@ -10,6 +10,7 @@ import 'package:logging/logging.dart';
 void main() {
   final loggingHttpClient = LoggingHttpClient();
 
+  Logger.root.level = Level.ALL;
   FlutterError.onError = (details) {
     Logger(
       'FlutterError',
@@ -32,8 +33,7 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  _MyAppState();
-
+  final navigatorKey = GlobalKey<NavigatorState>();
   late DebugPageController _debugPageController;
 
   @override
@@ -43,6 +43,7 @@ class _MyAppState extends State<MyApp> {
     _debugPageController = DebugPageController(
       showEntryButton: true,
       loggingHttpClient: widget._loggingHttpClient,
+      navigatorKey: navigatorKey,
     );
   }
 
@@ -52,6 +53,8 @@ class _MyAppState extends State<MyApp> {
       controller: _debugPageController,
       child: MaterialApp(
         title: 'Debug Page Demo',
+        navigatorKey: navigatorKey,
+        navigatorObservers: [_debugPageController.navigatorObserver],
         theme: ThemeData(
           colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
           useMaterial3: true,
@@ -133,6 +136,9 @@ class _MyHomePageState extends State<MyHomePage> {
   void _generateTestLogs() {
     // Generate various types of logs to demonstrate the enhanced log details
 
+    // Simple debug log
+    _logger.fine("That's a fine message!");
+
     // Simple info log
     _logger.info('This is a simple info message');
 
@@ -189,25 +195,35 @@ class _MyHomePageState extends State<MyHomePage> {
               const SizedBox(height: 16),
               ElevatedButton(
                 onPressed: () {
-                  Overlay.of(context).insert(
-                    OverlayEntry(
-                      opaque: true,
-                      builder: (context) => Container(
-                        color: Colors.red,
-                        alignment: Alignment.center,
-                        child: const Material(
-                          color: Colors.transparent,
-                          child: Text(
-                            'Overlay',
-                            style: TextStyle(
-                              color: Colors.white,
-                              fontSize: 24,
+                  late OverlayEntry entry;
+                  entry = OverlayEntry(
+                    opaque: true,
+                    builder: (context) => Container(
+                      color: Colors.red,
+                      alignment: Alignment.center,
+                      child: Material(
+                        color: Colors.transparent,
+                        child: Column(
+                          mainAxisSize: MainAxisSize.min,
+                          children: [
+                            const Text(
+                              'Overlay',
+                              style: TextStyle(
+                                color: Colors.white,
+                                fontSize: 24,
+                              ),
                             ),
-                          ),
+                            ElevatedButton(
+                              onPressed: () => entry.remove(),
+                              child: const Text('Close overlay'),
+                            ),
+                          ],
                         ),
                       ),
                     ),
                   );
+
+                  Overlay.of(context).insert(entry);
                 },
                 child: const Text('Show an overlay'),
               ),
@@ -217,6 +233,7 @@ class _MyHomePageState extends State<MyHomePage> {
       ),
       floatingActionButtonLocation: FloatingActionButtonLocation.startFloat,
       floatingActionButton: FloatingActionButton.extended(
+        heroTag: 'Send request',
         label: const Text('Send request'),
         icon: const Icon(Icons.send),
         tooltip: 'Send a request',