+++ packages/leancode_debug_page/lib/src/core/debug_page_controller.dart
@@ -1,11 +1,14 @@
 import 'dart:async';
 
 import 'package:flutter/foundation.dart';
+import 'package:flutter/widgets.dart';
 import 'package:leancode_debug_page/src/core/logger_listener.dart';
 import 'package:leancode_debug_page/src/core/logging_http_client.dart';
 import 'package:leancode_debug_page/src/core/shake_detector.dart';
 import 'package:leancode_debug_page/src/models/filter.dart';
 import 'package:leancode_debug_page/src/models/request_log_record.dart';
+import 'package:leancode_debug_page/src/ui/debug_page_route.dart';
+import 'package:leancode_debug_page/src/ui/logs_inspector/logs_inspector.dart';
 import 'package:logging/logging.dart';
 import 'package:rxdart/rxdart.dart';
 
@@ -15,6 +18,7 @@ class DebugPageController {
     this.showOnShake = true,
     this.ignoredBasePath,
     required this.loggingHttpClient,
+    required this.navigatorKey,
   }) : loggerListener = LoggerListener() {
     requestsFilters = ValueNotifier([]);
     _sourceRequestsStreamSubscription = loggingHttpClient.logStream.listen(
@@ -37,10 +41,15 @@ class DebugPageController {
     if (showOnShake) {
       _shakeDetector = ShakeDetector.autoStart(
         shakeThresholdGravity: 4,
-        onPhoneShake: () => visible.value = true,
+        onPhoneShake: open,
       );
     }
   }
+
+  late final NavigatorObserver navigatorObserver = _Observer(_isOpen);
+
+  final _isOpen = ValueNotifier(false);
+  ValueListenable<bool> get isOpen => _isOpen;
 
   final bool showEntryButton;
   final bool showOnShake;
@@ -49,10 +58,9 @@ class DebugPageController {
 
   final LoggingHttpClient loggingHttpClient;
   final LoggerListener loggerListener;
+  final GlobalKey<NavigatorState> navigatorKey;
 
   ShakeDetector? _shakeDetector;
-
-  final visible = ValueNotifier(false);
 
   late ValueNotifier<List<Filter<RequestLogRecord>>> requestsFilters;
   late ValueNotifier<List<Filter<LogRecord>>> loggerFilters;
@@ -79,17 +87,48 @@ class DebugPageController {
     _loggerLogController.add(filtered);
   }
 
+  void open() => navigatorKey.currentState!.push(LogsInspectorRoute(this));
+
+  void close() =>
+      navigatorKey.currentState!.popUntil((route) => route is! DebugPageRoute);
+
   void dispose() {
     loggerListener.dispose();
 
     _shakeDetector?.stopListening();
-
-    visible.dispose();
 
     _sourceRequestsStreamSubscription.cancel();
     requestsFilters.dispose();
 
     _sourceLoggerLogsStreamSubscription.cancel();
     loggerFilters.dispose();
+  }
+}
+
+class _Observer extends NavigatorObserver {
+  _Observer(this.notifier);
+
+  final ValueNotifier<bool> notifier;
+
+  final _debugPageRoutes = <DebugPageRoute>[];
+
+  void _updateNotifier() {
+    notifier.value = _debugPageRoutes.isNotEmpty;
+  }
+
+  @override
+  void didPush(Route<dynamic> route, Route<dynamic>? previousRoute) {
+    if (route is DebugPageRoute) {
+      _debugPageRoutes.add(route);
+      _updateNotifier();
+    }
+  }
+
+  @override
+  void didPop(Route<dynamic> route, Route<dynamic>? previousRoute) {
+    if (route is DebugPageRoute) {
+      _debugPageRoutes.remove(route);
+      _updateNotifier();
+    }
   }
 }