+++ packages/leancode_debug_page/lib/src/ui/debug_page_overlay.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:leancode_debug_page/leancode_debug_page.dart';
 import 'package:leancode_debug_page/src/ui/debug_page_button.dart';
-import 'package:leancode_debug_page/src/ui/logs_inspector/logs_inspector.dart';
 
 class DebugPageOverlay extends StatelessWidget {
   const DebugPageOverlay({
@@ -23,27 +22,16 @@ class DebugPageOverlay extends StatelessWidget {
       fit: StackFit.passthrough,
       children: [
         child,
-        if (_controller.showEntryButton)
-          DebugPageButton(controller: _controller),
         ValueListenableBuilder(
-          valueListenable: _controller.visible,
-          builder: (context, visible, child) {
-            if (!visible || child == null) {
-              return const SizedBox();
+          valueListenable: _controller.isOpen,
+          builder: (context, isOpen, child) {
+            if (_controller.showEntryButton && !isOpen) {
+              return child!;
             }
 
-            return child;
+            return const SizedBox();
           },
-          child: MaterialApp(
-            theme: ThemeData(
-              useMaterial3: true,
-              colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
-            ),
-            home: LogsInspector(
-              controller: _controller,
-              onBackButtonClicked: () => _controller.visible.value = false,
-            ),
-          ),
+          child: DebugPageButton(controller: _controller),
         ),
       ],
     );