+++ packages/leancode_debug_page/lib/src/ui/logs_inspector/requests/requests_tab_filters_menu.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:leancode_debug_page/leancode_debug_page.dart';
 import 'package:leancode_debug_page/src/core/filters/request_filters.dart';
 import 'package:leancode_debug_page/src/models/request_log_record.dart';
-import 'package:leancode_debug_page/src/ui/colors.dart';
 import 'package:leancode_debug_page/src/ui/logs_inspector/widgets/filtered_search_field.dart';
 import 'package:leancode_debug_page/src/ui/logs_inspector/widgets/labeled_dropdown.dart';
 import 'package:leancode_debug_page/src/ui/typography.dart';
@@ -68,8 +67,7 @@ class _RequestsTabFiltersMenuState extends State<RequestsTabFiltersMenu> {
   Widget build(BuildContext context) {
     return Container(
       padding: const EdgeInsets.fromLTRB(24, 0, 24, 16),
-      color: DebugPageColors.background,
-      child: DefaultTextStyle(
+      child: DefaultTextStyle.merge(
         style: DebugPageTypography.medium,
         child: Column(
           mainAxisSize: MainAxisSize.min,